+++ algolia/internal/opt/optional_filters_test.go
@@ -117,19 +117,39 @@ func TestOptionalFilters_LegacyDeserialization(t *testing.T) {
 		},
 		{
 			`"filter1:value1,filter2:value2"`,
-			opt.OptionalFilterOr("filter1:value1", "filter2:value2"),
+			opt.OptionalFilterAnd("filter1:value1", "filter2:value2"),
 		},
 		{
 			`" filter1:value1 , filter2:value2 "`,
+			opt.OptionalFilterAnd("filter1:value1", "filter2:value2"),
+		},
+		{
+			`"(filter1:value1,filter2:value2)"`,
 			opt.OptionalFilterOr("filter1:value1", "filter2:value2"),
 		},
 		{
 			`["filter1:value1","filter2:value2"]`,
-			opt.OptionalFilterOr("filter1:value1", "filter2:value2"),
+			opt.OptionalFilterAnd("filter1:value1", "filter2:value2"),
 		},
 		{
 			`[" filter1:value1 "," filter2:value2 "]`,
+			opt.OptionalFilterAnd("filter1:value1", "filter2:value2"),
+		},
+		{
+			`[["filter1:value1","filter2:value2"]]`,
 			opt.OptionalFilterOr("filter1:value1", "filter2:value2"),
+		},
+		{
+			`[["filter1:value1","filter2:value2"], "filter3:value3"]`,
+			opt.OptionalFilterAnd(opt.OptionalFilterOr("filter1:value1", "filter2:value2"), "filter3:value3"),
+		},
+		{
+			`["filter1:value1,filter2:value2","filter3:value3"]`,
+			opt.OptionalFilterAnd("filter1:value1,filter2:value2", "filter3:value3"),
+		},
+		{
+			`"(filter1:value1,filter2:value2),filter3:value3"`,
+			opt.OptionalFilterAnd(opt.OptionalFilterOr("filter1:value1", "filter2:value2"), "filter3:value3"),
 		},
 	} {
 		var got opt.OptionalFiltersOption