+++ algolia/opt/composable_filter.go
@@ -2,6 +2,7 @@ package opt
 
 import (
 	"encoding/json"
+	"fmt"
 	"reflect"
 	"strings"
 
@@ -74,25 +75,63 @@ func (o *composableFilterOption) UnmarshalJSON(data []byte) error {
 	}
 
 	var (
-		ok     = false
-		filter string
-		ors    []string
-		ands   [][]string
+		ok      = false
+		filter  string
+		ors     []string
+		ands    [][]string
+		options []interface{}
 	)
 
+	// Handles legacy one-string format as filters.
+	// Adds outer groups as `ands`, adds inner groups as `ors` if there are any.
+	//"A:1,B:2"       => [["A:1"],["B:2"]]
+	//"(A:1,B:2),C:3" => [["A:1","B:2"],["C:3"]]
+	//"(A:1,B:2)"     => [["A:1","B:2"]]
 	if json.Unmarshal(data, &filter) == nil {
 		ok = true
-		ors = strings.Split(filter, ",")
+		replacer := strings.NewReplacer("(", " ", ")", " ")
+
+		var start, count int
+		for i, c := range filter {
+			switch c {
+			case '(':
+				count++
+			case ')':
+				count--
+			case ',':
+				if count == 0 {
+					// remove parentheses and split filters by comma
+					ors = strings.Split(replacer.Replace(filter[start:i]), ",")
+					ands = append(ands, ors)
+					start = i + 1
+				}
+			}
+		}
+
+		// add last chunk of the filter
+		ors = strings.Split(replacer.Replace(filter[start:]), ",")
 		ands = append(ands, ors)
 	}
 
-	if !ok && json.Unmarshal(data, &ors) == nil {
+	if !ok && json.Unmarshal(data, &options) == nil {
 		ok = true
-		ands = append(ands, ors)
-	}
 
-	if !ok && json.Unmarshal(data, &ands) == nil {
-		ok = true
+		for _, option := range options {
+			switch v := option.(type) {
+			case []interface{}:
+				ors = []string{}
+
+				for _, val := range v {
+					ors = append(ors, fmt.Sprint(val))
+				}
+
+				ands = append(ands, ors)
+			case string:
+				ands = append(ands, []string{v})
+			default:
+				return errs.ErrJSONDecode(data, "composableFilterOption (string or []string or [][]string)")
+			}
+		}
 	}
 
 	if !ok {